Removed old spec classes, So that new phpspec tests can be written implementing the behat tests that have been written
