+++ README.md
@@ -1,3 +1,5 @@
 # personService
 
-<p>A service that describes a person and allows them to add and remove their favorite foods.</p> 
+[![Build Status](https://travis-ci.org/nunchuck/personService.svg?branch=master)](https://travis-ci.org/nunchuck/personService)
+
+<p>A service that describes a person and allows them to add and remove their favorite foods.</p>
