+++ tasks.txt
@@ -1,0 +1,17 @@
+/*
+     * Tasks:
+     *
+     * - Complete Person class as above
+     * - Complete Employee class which inherits from Person
+     * - Get Person and Employee stored/retrieved in a DB (create DB first)
+     *      - DB Class (& using an Interface?)
+     * - Get this all working as a RESTful service over HTTP
+     *
+     *
+     * Things to read up on/think about:
+     *
+     * - Class Inheritance and Interfaces
+     * - Database design for Person and Employee
+     * - Microservices
+     * - REST/RESTful services (how you should interact with your service)
+     */
